Makefile: rewrite to do dependency tracking
The makefile will now check for Xilinx tools in environment and suggest
to the user that they either install ISE or source the setup script.

The list of HDL files are in a seperate file called hdl.mk that is
included in the top level Makefile. This makes the actual make rules
easier to read.

The gitignore is updated to ignore the generated files.

Note that the 'program' rule is not finished, and does not work.

Signed-off-by: Joel Stanley <joel@jms.id.au>